+++ src/app/events/events-routing.module.ts
@@ -5,11 +5,17 @@ import { EventsComponent } from '@app/events/events.component';
 import { EventViewComponent } from '@app/events/event-view/event-view.component';
 import { EventAdminComponent } from '@app/events/event-admin/event-admin.component';
 import { EventEditComponent } from '@app/events/event-edit/event-edit.component';
+import { EventCreationComponent } from '@app/events/event-creation/event-creation.component';
+
 
 const routes: Routes = [
   {
     path: '',
     component: EventsComponent
+  },
+  {
+    path: 'new',
+    component: EventCreationComponent,
   },
   {
     path: ':id',